refactor: move atomic components to UIBlocks dir
