+++ src/components/ReflectionCard.tsx
@@ -1,6 +1,6 @@
 import { type ReflectionsQuery } from "~/graphql/generated"
-import DoubleQuotesIcon from "../atoms/DoubleQuotesIcon"
-import ImageWithFallback from "../atoms/ImageWithFallback"
+import DoubleQuotesIcon from "~/components/UIBlocks/DoubleQuotesIcon"
+import ImageWithFallback from "~/components/UIBlocks/ImageWithFallback"
 
 type ReflectionCardProps = {
   reflection: ReflectionsQuery["reflections"][number]
