+++ src/pages/_app.tsx
@@ -1,5 +1,5 @@
 import { AppProps } from "next/app"
-import GraphQLClientProvider from "../lib/client"
+import GraphQLClientProvider from "~/lib/client"
 import "~/styles/globals.css"
 
 function MyApp({ Component, pageProps }: AppProps) {
